requirements.txt
langchain_openai
langchain_core
langchain_community
streamlit
langserve
fastapi
uvicorn
sse_starlette
langchain_ollama